Fixes enemy removal of head node.
previously, I implemented a poor design for removing nodes from a linked
list. Mainly, I had an entirely separate execution path for removing the head
node vs traversing the list. This was not only cumbersome to maintain,
read, and grok, but also infused some bugs into the system in which I
was unable to ever remove the first word.

This update builds a proper linked list traversal algorithm thus
removinng the separate execution path and the bug entirely.