Updated Makefile to fail if it can't find dictionary
The dictionary file can be named/located several different ways
depending on your OS. This commit changes it from being hard coded in
the *.c file, to being dynamically set on the commandline via some
`make` testing. I feel that this is a simple solution that provides
extensibility without actually adding to the complexity of the C code
itself. Plus it fails the build instead of allowing a build and the
application being the one to notify you if something went wrong.