Makefile now runs on MAC correctly
The makefile ifeq tree for finding the dictionary file wasn't working
correctly on MAC. I figured out that the original statement of `echo -n
yes`, which doesn't print a "newline" was causing the make to fail on
MAC. I added that for compatibilty... but I guess thats not the case
anymore.

This commit removes the `-n`, thus allowing `echo` to print the newline
characters; thus resulting in proper evaluation from `make`.